+++ workspaces/argo-workflows/plugins/argo-workflows/dev/index.tsx
@@ -15,14 +15,14 @@
  */
 
 import { Entity } from '@backstage/catalog-model';
-import { discoveryApiRef, fetchApiRef } from '@backstage/core-plugin-api';
+import { discoveryApiRef, fetchApiRef } from '@backstage/frontend-plugin-api';
 import { createDevApp } from '@backstage/dev-utils';
 import { EntityProvider } from '@backstage/plugin-catalog-react';
 import { TestApiProvider } from '@backstage/test-utils';
 import { FullPage, PluginHeader } from '@backstage/ui';
 import { RiFlowChart } from '@remixicon/react';
 
-import { argoWorkflowsPlugin, ArgoWorkflowsCI } from '../src/plugin';
+import { Router as ArgoWorkflowsCI } from '../src/components/Router';
 import {
   succeededWorkflow,
   runningWorkflow,
@@ -178,7 +178,6 @@ function DevPage({ entity, title }: { entity: Entity; title: string }) {
 }
 
 createDevApp()
-  .registerPlugin(argoWorkflowsPlugin)
   .addPage({
     element: <DevPage entity={multiInstanceEntity} title="Argo Workflows" />,
     title: 'Workflows',
