+++ workspaces/argo-workflows/plugins/argo-workflows/package.json
@@ -6,14 +6,10 @@
   "types": "src/index.ts",
   "exports": {
     ".": "./src/index.ts",
-    "./alpha": "./src/alpha.tsx",
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
-      "alpha": [
-        "src/alpha.tsx"
-      ],
       "package.json": [
         "package.json"
       ]
