+++ workspaces/argo-workflows/plugins/argo-workflows/src/index.ts
@@ -20,5 +20,8 @@
  * @packageDocumentation
  */
 
-export { argoWorkflowsPlugin, ArgoWorkflowsCI } from './plugin';
+// Old frontend system support
+export * from './legacy';
+
+export { default } from './plugin';
 export { isArgoWorkflowsAvailable } from '@backstage-community/plugin-argo-workflows-common';
