+++ workspaces/argo-workflows/plugins/argo-workflows/src/legacy.ts
@@ -14,31 +14,4 @@
  * limitations under the License.
  */
 
-import {
-  createPlugin,
-  createComponentExtension,
-} from '@backstage/core-plugin-api';
-
-/**
- * The Argo Workflows frontend plugin.
- *
- * @public
- */
-export const argoWorkflowsPlugin = createPlugin({
-  id: 'argo-workflows',
-});
-
-/**
- * Component extension that renders the Argo Workflows CI/CD view
- * for a catalog entity.
- *
- * @public
- */
-export const ArgoWorkflowsCI = argoWorkflowsPlugin.provide(
-  createComponentExtension({
-    name: 'ArgoWorkflowsCI',
-    component: {
-      lazy: () => import('./components/Router').then(m => m.Router),
-    },
-  }),
-);
+export * from './legacy-plugin';
